third_party/blink/web_tests/external/wpt/css/css-transitions/KeyframeEffect-getKeyframes.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] KeyframeEffect.getKeyframes() returns expected frames for a transition after resetting the effect target
  assert_equals: properties on ComputedKeyframe #0 expected "composite,computedOffset,easing,left,offset" but got "composite,computedOffset,easing,offset"
[FAIL] KeyframeEffect.getKeyframes() returns expected frames for a custom property transition
  Cannot read properties of undefined (reading 'effect')
Harness: the test ran to completion.
